src/ui/lib_analysis/db/requirements.txt
streamlit>=1.32
pandas>=2.0
psycopg[binary]>=3.1
psycopg2-binary>=2.9
